combined auth routes and middleware
