--- src/server/src/middleware/requireJwtAuth.js
@@ -1,5 +1,0 @@
-import passport from "passport";
-
-const requireJwtAuth = passport.authenticate("jwt", { session: false });
-
-export default requireJwtAuth;
