+++ src/server/src/routes/test.js
@@ -1,9 +1,6 @@
 import { Router } from "express";
-import authRoutes from "./auth/index.js";
 
 const router = Router();
-
-router.use("/auth", authRoutes);
 
 router.get("/", (req, res) => {
   res.json({ message: "Hello from the server!" });
